Huge tidy up of skins.
Now all skins are based on one skin (called weblearn) and custom ones redirect most of their requests back to the WebLearn one. When a skin has custom CSS it imports the weblearn.css skin and then overrides bits.

The redirect filter is used to redirect all requests that aren't part of the custom skin.